Added raw download technique of a single file.
I don't know how to set the MIME type if possible to the galaxy server.
But once I find that out I will be able to compress the file.
Stephen warns of making sure to use galaxy's own job queue and NOT to
go eat up a CPU core to do the compression work.